requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
authlib==1.3.1
starlette==0.38.5
stripe==10.14.0
SQLAlchemy==2.0.34
itsdangerous==2.2.0
orjson==3.10.7
